// src/app/models/b2b-user-model.ts
export class b2bUserModel{
    userId ?:       string;
    userEmail?:       string;
    userPass?:       string;
    userTlf?:       string;
    userName?:       string;
    userLastName?:       string;
    photoProfileUser?:       string;
    createAt?:       string;
    userAbout?:       string;
}